separate concerns of FileUtils/Serializable classes
